requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-multipart
pytest
httpx
jinja2
python-multipart
python-multipart>=0.0.9
psycopg[binary]
alembic
prometheus-client
passlib
PyJWT
pydantic-settings
bcrypt
